fetch/requirements.txt
certifi==2019.6.16
chardet==3.0.4
geojson==2.4.1
idna==2.8
overpass==0.6.1
pyaml==19.4.1
PyYAML==5.1.2
requests==2.22.0
Shapely==1.6.4.post2
urllib3==1.25.3
